backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic[email]>=2.6.3
pydantic-settings>=2.2.1
firebase-admin>=6.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
pandas>=2.2.1
scikit-learn>=1.4.1.post1
numpy>=1.26.4
python-dotenv>=1.0.1
httpx>=0.27.0
